# advertisementservice/src/main/resources/openapi/external.yaml
openapi: 3.0.0
...
info:
  title: external
  version: '1.0'
  contact:
    name: Balint Jeszenszky
  description: Classified advertisement site advertisement microservice external api
servers:
  - url: 'http://localhost:8082/api/advertisement/external'
paths:
  '/advertisement/{id}':
    parameters:
      - schema:
          type: integer
        name: id
        in: path
        required: true
    get:
      summary: Your GET endpoint
      tags:
        - advertisement
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdvertisementResponse'
        '404':
          description: Not Found
      operationId: get-advertisement-id
      description: Get a single advertisement by id
    put:
      summary: ''
      operationId: put-advertisement-id
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdvertisementResponse'
        '400':
          description: Bad Request
        '403':
          description: Forbidden
      description: Edit an advertisement by id
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/EditAdvertisementRequest'
      tags:
        - advertisement
    delete:
      summary: ''
      operationId: delete-advertisement-id
      responses:
        '204':
          description: No Content
        '403':
          description: Forbidden
      description: Delete an advertisement by id
      tags:
        - advertisement
...
components:
  schemas:
    AdvertisementResponse:
      title: AdvertisementResponse
      x-stoplight:
        id: v8km5zj6k0ca3
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        advertiserId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        price:
          type: number
        categoryId:
          type: integer
        status:
          type: string
          enum:
            - AVAILABLE
            - FREEZED
            - SOLD
            - ARCHIVED
            - BIDDING
        type:
          type: string
          enum:
            - FIXED_PRICE
            - BID
        expiration:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - advertiserId
        - createdAt
        - updatedAt
        - price
        - status
        - type
...
    EditAdvertisementRequest:
      title: EditAdvertisementRequest
      x-stoplight:
        id: pjkx34r6cyowt
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        categoryId:
          type: integer
        status:
          type: string
          enum:
            - AVAILABLE
            - FREEZED
            - SOLD
            - ARCHIVED
            - BIDDING
        images:
          type: array
          format: binary
          items:
            type: string
            format: binary
        deletedImages:
          type: string
      required:
        - title
        - description
        - price
        - categoryId
        - status
...
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
...
security:
  - JWT: []